+++ gatsby-node.js
@@ -3,12 +3,154 @@ const path = require('path')
 const { createFilePath } = require('gatsby-source-filesystem')
 const {cssModulesConfig} = require('gatsby-1-config-css-modules');
 
+/**
+ * createBlogPosts
+ * 
+ * @function creates pages for all blog posts
+ * @param {Array<Edge>} all the blog post edges.
+ * @param {Function} the createPage function
+ * @return {Array<Edge>} of posts
+ */
+
+const createBlogPosts = (posts, createPage) => {
+  posts.forEach(edge => {
+    const id = edge.node.id
+    createPage({
+      path: edge.node.fields.slug,
+      tags: edge.node.frontmatter.tags,
+      component: path.resolve(
+        `src/templates/${String(edge.node.frontmatter.templateKey)}.js`
+      ),
+      // additional data can be passed via context
+      context: {
+        id,
+      },
+    })
+  })
+
+  return posts;
+}
+
+const createCompanyPages = (companies, createPage) => {
+  companies.forEach(edge => {
+    const companyId = edge.node.companyId;
+
+    // TODO: Ensure that none of the company names are taken. They need to be unique,
+    // or we use the company handle.
+    
+    createPage({
+      path: edge.node.fields.slug,
+      component: path.resolve(`src/templates/company.js`),
+      context: {
+        companyId
+      }
+    })
+  })
+}
+
+/**
+ * createTagPages
+ * 
+ * @function creates all tag pages
+ * @param Array<String> of tags
+ * @param Function to create page
+ * @return Array<String>
+ */
+
+const createTagPages = (tags, createPage) => {
+  return tags.forEach(tag => {
+    const tagPath = `/tags/${_.kebabCase(tag)}/`
+
+    createPage({
+      path: tagPath,
+      component: path.resolve(`src/templates/tags.js`),
+      context: {
+        tag,
+      },
+    })
+  })
+}
+
+/**
+ * createCategoryPages
+ * 
+ * @function creates all category pages
+ * @param Array<String> of categories
+ * @param Function to create pages
+ * @return Array<String>
+ */
+
+const createCategoryPages = (categories, createPage) => {
+  return categories.forEach(category => {
+    const categoryPath = `/blog/categories/${_.kebabCase(category)}/`
+    
+    createPage({
+      path: categoryPath,
+      component: path.resolve('src/templates/category.js'),
+      context: {
+        category
+      }
+    })
+  })
+}
+
+/**
+ * getAllCategoriesFromPosts
+ * 
+ * @function that gets all categories from the post nodes
+ * and returns an array of unique categories
+ * @param {Array<Edge>} posts
+ * @return {Array<String} of categories
+ */
+
+const getAllCategoriesFromPosts = (posts) => {
+  let categories = [];
+  posts.forEach(edge => {
+    if (_.get(edge, `node.frontmatter.category`)) {
+      categories.push(edge.node.frontmatter.category)
+    }
+  })
+  return _.uniq(categories);
+}
+
+/**
+ * getAllTagsFromPosts
+ * 
+ * @function that gets all the tags from the post nodes
+ */
+
+const getAllTagsFromPosts = (posts) => {
+  let tags = [];
+  posts.forEach(edge => {
+    if (_.get(edge, `node.frontmatter.tags`)) {
+      tags = tags.concat(edge.node.frontmatter.tags)
+    }
+  })
+  return _.uniq(tags);
+}
+
+
+
+// exports.sourceNodes = async ({ boundActionCreators }) => {
+//   const { createNode } = boundActionCreators;
+//   // Create nodes here, generally by downloading data
+//   // from a remote API.
+//   // const data = await fetch(REMOTE_API);
+//   // const data = await dataSource.getData()
+
+//   // Process data into nodes.
+//   data.forEach(datum => createNode(dataSource.processData(datum)));
+
+//   // We're done, return.
+//   return;
+// };
+
 exports.createPages = ({ boundActionCreators, graphql }) => {
-  const { createPage } = boundActionCreators
+  const { createPage } = boundActionCreators;
 
   return graphql(`
     {
-      allMarkdownRemark(limit: 1000) {
+      posts: allMarkdownRemark(limit: 1000) {
         edges {
           node {
             id
@@ -23,6 +165,22 @@ exports.createPages = ({ boundActionCreators, graphql }) => {
           }
         }
       }
+
+      companies: allCompany(limit: 1000) {
+        edges {
+          node {
+            companyName
+            companyId
+            industry
+            logoUrl
+            earlyAdopter
+            brandImageUrl
+            fields {
+              slug
+            }
+          }
+        }
+      }
     }
   `).then(result => {
 
@@ -31,74 +189,19 @@ exports.createPages = ({ boundActionCreators, graphql }) => {
       return Promise.reject(result.errors)
     }
 
-    const posts = result.data.allMarkdownRemark.edges
+    // Create all blog posts
+    const posts = createBlogPosts(result.data.posts.edges, createPage);
 
-    posts.forEach(edge => {
-      const id = edge.node.id
-      createPage({
-        path: edge.node.fields.slug,
-        tags: edge.node.frontmatter.tags,
-        component: path.resolve(
-          `src/templates/${String(edge.node.frontmatter.templateKey)}.js`
-        ),
-        // additional data can be passed via context
-        context: {
-          id,
-        },
-      })
-    })
+    // Get tags and categories so that we can create pages
+    const tags  = getAllTagsFromPosts(posts);
+    const categories = getAllCategoriesFromPosts(posts);
 
-    // Tag pages:
-    let tags = []
-    let categories = [];
+    // Create tag and categories pages
+    createTagPages(tags, createPage);
+    createCategoryPages(categories, createPage);
 
-    // Iterate through each post, putting all found tags into `tags`
-    posts.forEach(edge => {
-      if (_.get(edge, `node.frontmatter.tags`)) {
-        tags = tags.concat(edge.node.frontmatter.tags)
-
-
-        /**
-         * Additionally, add all categories to an array
-         */
-
-        let category = edge.node.frontmatter.category
-
-        if (category) {
-          categories.push(category)
-        }
-        
-      }
-    })
-    // Eliminate duplicate tags
-    tags = _.uniq(tags)
-    categories = _.uniq(categories)
-
-    // Make tag pages
-    tags.forEach(tag => {
-      const tagPath = `/tags/${_.kebabCase(tag)}/`
-
-      createPage({
-        path: tagPath,
-        component: path.resolve(`src/templates/tags.js`),
-        context: {
-          tag,
-        },
-      })
-    })
-
-    // Make categories pages
-    categories.forEach(category => {
-      const categoryPath = `/blog/categories/${_.kebabCase(category)}/`
-      
-      createPage({
-        path: categoryPath,
-        component: path.resolve('src/templates/category.js'),
-        context: {
-          category
-        }
-      })
-    })
+    // Create pages from companies
+    createCompanyPages(result.data.companies.edges, createPage)
 
   })
 }
@@ -129,6 +232,14 @@ exports.onCreateNode = ({ node, boundActionCreators, getNode }) => {
       name: `slug`,
       node,
       value,
+    })
+  }
+
+  else if (node.internal.type === 'Company') {
+    createNodeField({
+      name: 'slug',
+      node,
+      value: `/companies/${_.kebabCase(node.companyName)}/`
     })
   }
 }